requirements.txt
torch
transformers
Pillow
einops
accelerate